requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.3.0
yarl==1.8.1
ymaps==1.1
